+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - `#58`, as user, i want to optionally use contraction 70.
 - `#59`, as user, i want to optionally use contraction 80.
 - `#60`, as user, i want to optionally use contraction 90.
+- `#61`, as user, i want to optionally use contraction 200.
 
 ## `1.8.0` - 5 July 2022